- Added Release configuration - Lot of internal modifications, adds and bug corrections - Updated for lastest Dofus version - Now, let's go for functionalities! :) => BEWARE: SQL files aren't up to date

+++ Game/Character.h
@@ -150,7 +150,7 @@ public:
 	CharacterStats& GetStats()
 	{ return m_stats; }
 
-	int GetLevel() const
+	uint8 GetLevel() const
 	{ return m_level; }
 
 	int8 GetBreed() const
